PODS-2433: code refactor files moved from vary to register
